API documentation

The basic Type

The starting point for EditorsRepo.jl is a directory in a local file system that includes subdirectories for standard kinds of content: editions of texts, indexing of texts to citable images and surfaces, and configuration files.

EditorsRepo.EditingRepository — Type

File organization of a repository for HMT editors.

Members

editions Path to directory with source files for textual editions.

dse Path to directory with delimited text files recording text, image and surface relations.

configs Path to directory with catalog.cex and citation.cex files.

EditorsRepo.repository — Function

Construct an EditingRepository from a root directory containing all three required directories.

repository(repodir; editions, dse, config)

If the required directories are named "editions", "dse" and "config", repodir is the only parameter needed. Optionally, you may supply names for any of those directories relative to reporoot.

EditorsRepo.configdir — Function

Find configuration directory for repository.

configdir(repo)

EditorsRepo.dsedir — Function

Find DSE directory for repository.

dsedir(repo)

EditorsRepo.editionsdir — Function

Find editions directory for repository.

editionsdir(repo)

Text configuration:

EditorsRepo.citationconfig — Function

Read citation configuration into a Table.

citationconfig(repo; filename, delimiter)

EditorsRepo.filename — Function

filename(repo, txturn)

Lookup file name in a repository for a text identified by URN.

Missing docstring.

Missing docstring for o2converter, diplomaticbuilder. Check Documenter's build log for details.

Missing docstring.

Missing docstring for normalizedbuilder. Check Documenter's build log for details.

EditorsRepo.orthography — Function

orthography(repo, txturn)

Lookup orthography for a text identified by URN.

Text catalog and contents:

EditorsRepo.textcatalog — Function

Read a TextCatalogCollection for a repository.

textcatalog(repo; catalogname, delimiter)

EditorsRepo.texturns — Function

Find CTS URNs of all texts cataloged as online.

texturns(repo)

EditorsRepo.archivalcorpus — Function

Create a citable corpus of all archival text in a repo.

archivalcorpus(r)

EditorsRepo.diplomaticcorpus — Function

Create diplomatic editions of all texts in a repo.

diplomaticcorpus(r)

EditorsRepo.normalizedcorpus — Function

Create normalized editions of all texts in a repo.

normalizedcorpus(r)

Classified tokenization:

Missing docstring.

Missing docstring for tokencorpus. Check Documenter's build log for details.

DSE records:

EditorsRepo.dsetriples — Function

dsetriples(repo)

List .cex files in DSE directory.

EditorsRepo.passageurnsforsurface — Function

passageurnsforsurface(r, u)

Compute list of passages in DSE records for a given surface.

EditorsRepo.diplomaticforsurface — Function

diplomaticforsurface(r, u)

Assemble citable passages in diplomatic edition for a given surface.

EditorsRepo.imagesforpassage — Function

imagesforpassage(r, u)

Compute list of images in DSE records for a given passage.

HTML utilities:

EditorsRepo.htmltoken — Function

Format an orthographic token in HTML.

htmltoken(psg, tokentype, ok)

Adds appropriate whitespace to non-punctuation tokens, and wraps orthographically failed tokens in span element.

Internal functions

EditorsRepo.xmlfiles — Function

xmlfiles(repository)

Find names of XML files in editing directory.

EditorsRepo.textpassages — Function

Select from a corpus those matching a given URN, omitting ref passages conventionally used for non-textual metadata.

textpassages(c, u)

EditorsRepo.isref — Function

True if last component of CTS URN passage is "ref". MID convention is to exclude elements, like notes on HMT scholia, with this identifier.

isref(urn)

EditorsRepo.passages_for_tokens — Function

Create a list of CitablePassages from a list of OrthographicTokens and a passage URN.

passages_for_tokens(tkns, urn)

Passage URNs are extended with an additional level of citation for the individual token. This citation tier is made up of sequential numbers for lexical tokens, and token number + a character for other kinds of tokens (1a, 1b, etc.).

EditorsRepo.analyzedtokens — Function

Compute a list of OrthographicTokens based on the normalized edition of all texts in the repository. The result is a list of tuples pairing a citable text passage for the token and its type.

analyzedtokens(r)

For a single token citable as a CitablePassage, find its token class and determine if it is orthographically valid.

analyzedtokens(r, cn)

Returns a tuple of a token type and a boolean value.