Indexing texts, images and surfaces (DSE records)

Exported:

EditorsRepo.dsefiles — Function

dsefiles(repository)

List .cex files in DSE directory.

EditorsRepo.dse_df — Function

dse_df(repository)

Merge all DSE data into a single dataframe.

EditorsRepo.surfaces — Function

surfaces(repo)

Compute list of unique surfaces in DSE records.

EditorsRepo.passages — Function

passages(repo)

Compute list of passages in DSE records.

EditorsRepo.images — Function

images(repo)

Compute list of passages in DSE records.

EditorsRepo.surfaceDse — Function

Find DSE records for surface currently selected in popup menu.

surfaceDse(repo, surfurn)

Internal:

EditorsRepo.baseurn — Function

Find URN for a single passage from DSE record, which could include a range with subrefs within a single passage.

baseurn(urn)