Text configuration

Loading and verifying the text configuration

EditorsRepo.citation_df — Function

Read citation configuration into a DataFrame.

citation_df(repo)

EditorsRepo.textcatalog — Function

Loads catalog of texts from file as an Array of CSV Rows.

textcatalog(repo, catalogname)
textcatalog(repo, catalogname, delimiter)

Arguments

repo The editorial repository.

catalogname Name of the text catalog file.

delimiter Delimiting string in the text catalog.

EditorsRepo.textcatalog_df — Function

Load catalog of texts from file as a DataFrame

textcatalog_df(repo)

Arguments

repo The editorial repository.

EditorsRepo.texturns — Function

Find CTS URNs of all texts cataloged as online.

texturns(repo)

EditorsRepo.missingcitation — Function

missingcitation(repo)

List entries with missing values in citation configuration.

EditorsRepo.filesmatch — Function

True if files on disk match listing in citation configuration.

EditorsRepo.catalogonly — Function

List texts in catalog but missing from citation config.

EditorsRepo.citedonly — Function

List configured files not found in file system.

Reading configuration settings

EditorsRepo.filename — Function

filename(repo, u)

Lookup file name in a repository for a text identified by URN.

filename(df, u)

Lookup file name in a DataFame for a text identified by URN.

EditorsRepo.o2converter — Function

o2converter(repo, u)

Lookup ohco2 converter in a repository for a text identified by URN.

o2converter(df, u)

Lookup ohco2 converter in a DataFrame for a text identified by URN.

EditorsRepo.diplomaticbuilder — Function

diplomaticbuilder(repo, u)

Lookup in a repository the builder for diplomatic edition for a text identified by URN.

diplomaticbuilder(df, u)

Lookup in a DataFrame the builder for diplomatic edition for a text identified by URN.

EditorsRepo.normalizedbuilder — Function

normalizedbuilder(repo, u)

Lookup in a repository the builder for diplomatic edition for a text identified by URN.

Lookup in a DataFrame the builder for diplomatic edition for a text identified by URN.

normalizedbuilder(df, u)

EditorsRepo.orthography — Function

Lookup in a repository the orthographic system for a text identified by URN.

orthography(repo, u)

orthography(df, u)

Lookup in a DataFrame the orthographic system for a text identified by URN.

EditorsRepo.ohco2forurn — Function

Eval the string expression for the OHCO2 converter configured for a URN.

ohco2forurn(textconfig, urn)

EditorsRepo.orthographyforurn — Function

Eval the string expression for the orthography system configured for a URN.

orthographyforurn(textconfig, urn)

EditorsRepo.diplomaticforurn — Function

Eval the string expression for the diplomatic builder configured for a URN.

diplomaticforurn(textconfig, urn)

Arguments

textconfig configuration file

urn text to retrieve

EditorsRepo.normalizerforurn — Function

Eval the string expression for the normalized builder configured for a URN.