Reading texts from the repository

Finding editions

EditorsRepo.xmlfiles — Function

xmlfiles(repository)

Find names of XML files in editing directory.

EditorsRepo.xmlfiles_df — Function

xmlfiles_df(repository)

Find names of XML files in editing directory.

Reading text contents

Exported functions:

EditorsRepo.diplomaticnodes — Function

Compose an array of CitableNodes for a diplomatic reading of a text identified by CtsUrn."

diplomaticnodes(repo, urn)

EditorsRepo.normalizednodes — Function

Compose an array of CitableNodes for a normalized reading of a text identified by CtsUrn.

normalizednodes(repo, urn)

EditorsRepo.diplnodetext — Function

Collect diplomatic text for a single text passage identified by URN. The URN should either match a citable node, or be a containing node for one or more citable nodes. Ranges URNs are not supported.

diplnodetext(repo, urn)

EditorsRepo.normednodetext — Function

Collect diplomatic text for a text passage identified by URN. The URN should either match a citable node, or be a containing node for one or more citable nodes. Ranges URNs are not supported.

normednodetext(repo, urn)

EditorsRepo.textpassages — Function

Select from a list passages those URN matching a given URN, and omit "ref" passages conventionally used for non-text content.

textpassages(psgs, urn)

EditorsRepo.normedtokens — Function

Compose an edition of normalized tokens for allt exts in repository.

normedtokens(repo)

Compose an edition of normalized tokens matching urn.

normedtokens(repo, urn)

EditorsRepo.lextokens — Function

Compose an edition of lexical tokens matching urn.

lextokens(repo, urn)

Internal functions:

EditorsRepo.isref — Function

True if last component of CTS URN passage is "ref". MID convention is to exclude elements, like notes on HMT scholia, with this identifier.

isref(urn)

EditorsRepo.nodetext — Function

Collect text from a list of passages for a text passage identified by URN. The URN should either match a citable node, or be a containing node for one or more citable nodes. Ranges URNs are not supported.

EditorsRepo.nodesfortokens — Function

Create a list of CitableNodes from a list of OrthographicTokens and a node URN.

nodesfortokens(tkns, urn)

Node URNs are extended with an additional level of citation for the individual token. This citation tier is made up of sequential numbers for lexical tokens, and token number + a character for other kinds of tokens (1a, 1b, etc.).

EditorsRepo.lexnodesfortokens — Function

Create a list of CitableNodes for all lexical tokens in a list of OrthographicTokens for a given URN.

lexnodesfortokens(tkns, urn)

Node URNs are extended with an additional level of citation for the individual token. This citation tier is made up of sequential numbers for lexical tokens, and token number + a character for other kinds of tokens (1a, 1b, etc.).