API documentation

The basic Type

The starting point for EditorsRepo.jl is a directory in a local file system that includes subdirectories for standard kinds of content: editions of texts, indexing of texts to citable images and surfaces, and configuration files.

EditorsRepo.EditingRepository — Type

File organization of a repository for HMT editors.

Members

editions Full path to directory with source files for textual editions.

dse Full path to directory with delimited text files recording text, image and surface relations.

configs Full path to directory with catalog.cex and citation.cex files.