API documentation

The basic Type

The starting point for EditorsRepo.jl is a directory in a local file system that includes subdirectories for standard kinds of content: editions of texts, indexing of texts to citable images and surfaces, and configuration files.

EditorsRepo.EditingRepository — Type

File organization of a repository for HMT editors.

Extracting high-level data from the repository

The text catalog

EditorsRepo.citation_df — Function

citation_df(repo)

Read citation configuration into a DataFrame.

EditorsRepo.textcatalog — Function

Loads catalog of texts from file as an Array of CSV Rows.

textcatalog(repo, catalogname)
textcatalog(repo, catalogname, delimiter)

Arguments

repo The editorial repository.

catalogname Name of the text catalog file.

delimiter Delimiting string in the text catalog.

EditorsRepo.textcatalog_df — Function

Loads catalog of texts from file as a DataFrame

textcatalog_df(repo)

Arguments

repo The editorial repository.

EditorsRepo.texturns — Function

Find CTS URNs of all texts cataloged as online.

texturns(repo)

Functions

EditorsRepo.missingcitation — Method

missingcitation(repo)

List entries with missing values in citation configuration.

EditorsRepo.filename — Method

filename(df, u)

Lookup file name in a DataFame for a text identified by URN.

EditorsRepo.o2converter — Method

o2converter(df, u)

Lookup ohco2 converter in a DataFrame for a text identified by URN.

EditorsRepo.diplomaticbuilder — Method

diplomaticbuilder(df, u)

Lookup in a DataFrame the builder for diplomatic edition for a text identified by URN.

EditorsRepo.normalizedbuilder — Method

Lookup in a DataFrame the builder for diplomatic edition for a text identified by URN.

normalizedbuilder(df, u)

EditorsRepo.orthography — Method

orthography(df, u)

Lookup in a DataFrame the orthographic system for a text identified by URN.

EditorsRepo.textforurn — Method

Read the text contents of the online file identified by urn.

EditorsRepo.ohco2forurn — Method

Eval the string expression for the OHCO2 converter configured for a URN.

EditorsRepo.orthographyforurn — Method

Eval the string expression for the orthography system configured for a URN.

EditorsRepo.diplomaticforurn — Method

Eval the string expression for the diplomatic builder configured for a URN.

diplomaticforurn(textconfig, urn)

Arguments

textconfig configuration file

urn text to retrieve

EditorsRepo.normalizerforurn — Method

Eval the string expression for the normalized builder configured for a URN.

EditorsRepo.filesmatch — Method

True if files on disk match listing in citation configuration.

EditorsRepo.catalogonly — Method

List texts in catalog but missing from citation config.

EditorsRepo.citedonly — Method

List configured files not found in file system.

EditorsRepo.xmlfiles — Method

xmlfiles(repository)

Find names of XML files in editing directory.

EditorsRepo.xmlfiles_df — Method

xmlfiles_df(repository)

Find names of XML files in editing directory.

EditorsRepo.diplomaticnodes — Method

Compose an array of CitableNodes for a diplomatic reading of a text identified by CtsUrn.

EditorsRepo.normalizednodes — Method

Compose an array of CitableNodes for a normalized reading of a text identified by CtsUrn.

EditorsRepo.dsefiles — Method

dsefiles(repository)

List .cex files in DSE directory.

EditorsRepo.dse_df — Method

dse_df(repository)

Merge all DSE data into a single dataframe.

EditorsRepo.surfaces — Method

surfaces(repo)

Compute list of unique surfaces in DSE records.

EditorsRepo.passages — Method

passages(repo)

Compute list of passages in DSE records.

EditorsRepo.images — Method

images(repo)

Compute list of passages in DSE records.