Reading texts from the repository

Finding editions

EditorsRepo.xmlfiles — Function

xmlfiles(repository)

Find names of XML files in editing directory.

EditorsRepo.xmlfiles_df — Function

xmlfiles_df(repository)

Find names of XML files in editing directory.

Reading text contents

Missing docstring.

Missing docstring for textforurn. Check Documenter's build log for details.

EditorsRepo.diplomaticnodes — Function

Compose an array of CitableNodes for a diplomatic reading of a text identified by CtsUrn."

diplomaticnodes(repo, urn)

EditorsRepo.normalizednodes — Function

Compose an array of CitableNodes for a normalized reading of a text identified by CtsUrn.

normalizednodes(repo, urn)